Doctor_Online/requirements.txt
appdirs==1.4.4
asgiref==3.3.0
beautifulsoup4==4.9.3
boto3==1.17.50
botocore==1.20.50
certifi==2020.11.8
cffi==1.14.5
chardet==3.0.4
comtypes==1.1.8
cryptography==3.4.4
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.3
django-allauth==0.44.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.3.1
django-ckeditor==6.0.0
django-countries==7.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
django-filter==2.4.0
django-forms-bootstrap==3.1.0
django-js-asset==1.2.2
django-markdown==0.8.4
django-markdown-deux==1.0.5
django-model-utils==4.1.1
django-pagedown==2.2.0
django-phone-field==1.8.1
django-taggit==1.3.0
django-tempus-dominus==5.1.2.13
django-tinymce==3.2.0
djangoajax==3.2
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
docopt==0.6.2
env==0.1.0
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
Js2Py==0.70
Markdown==3.3.3
markdown2==2.4.0
oauthlib==3.1.0
packaging==20.8
Pillow==8.0.1
pipwin==0.5.0
psycopg2==2.8.6
PyArabic==0.6.10
PyAudio @ file:///C:/Users/abdal/AppData/Local/Programs/Python/Python38-32/Scripts/PyAudio-0.2.11-cp38-cp38-win32.whl
pycparser==2.20
pyjsparser==2.7.1
PyJWT==1.7.1
pyparsing==2.4.7
PyPDF2==1.26.0
pypiwin32==223
PyPrind==2.11.2
pySmartDL==1.3.4
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pyttsx3==2.90
pytz==2020.4
pywin32==300
requests==2.25.0
requests-oauthlib==1.3.0
s3transfer==0.3.6
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
stripe==2.55.2
swapper==1.1.2.post1
tqdm==4.56.2
tzlocal==2.1
urllib3==1.26.2
virtualenv==20.1.0
youtube-dl==2020.12.31
